feat(chat): refactor schema components
